mb done: the proof of that tiling can be secure
(and the conditions where it's secure).

[2:12 PM] Max: tl;dr: I think tiling requires that the chain-work ratio of various tiles obeys some rules:
* for the root tile: it has to be no more than $v$ times as secure as it's child tiles (otherwise attacker can 'sever' link to child tiles)
* for all other tiles: must have chain-work less than $r$ times that of the parent tile, where $r > \frac{1}{2}(\sqrt{4v-3} + 1)$
* so, network wide: $v > r > 1/2 * (\sqrt(4v-3)+1)$ -- for v=3: $3 > r > 2$

so, basically, provided we can maintain that: enforcing (for all tiles except leaves) $w_{p} > r * w_{p|i} is the security criterion for a tiling being secure. (where w_{p} is the chain-work of a parent tile, and w_{p|i} is the chain-work of one of that parent tile's children.) The only exception is the root tile, which doesn't have a parent, and we need to encforce w_{1} < v * w_{1|i} which implies r * w_{1|i} < v * w_{1|i} (nb: w_{1} is the chain-work of the root tile).
[2:15 PM] Max: how do we enforce those rules? by making sure that the chain-work contrib of tiles over the tiling have the right security distributions. but security/chain-work will be decided by block rewards, which will be decided by the distribution of ROO. This can't fail just b/c someone moves a bunch of ROO out of the root tile -- so we'll need a Work Distribution Adjustment Algorithm or something. Basically: if the root tile gets too big, then the block rewards of its children are artificially increased (and the root tile's block rewards decreased) so that miners move over. This is implemented in the conversion algorithm (in the fork rule).
[2:16 PM] Max: Similarly, the WDAA might artificially increase the block rewards of a parent tile if some tile get's too much ROO on it.
[2:19 PM] Max: The downside of tiling is that big doublespend attempts take depth many "tile confirmations" to resolve (which always resolves in the honest network's favor for attacks <51% of total hash-rate).

what's a "tile confirmation" something like: all simplex-chains in one tile reflecting all simplex-chains in the other --- so basically it operates at B_f -- the block frequency. If all simplex-chains have B_f = 1/15, then each tile has a confirmation rate of like 10-100 Hz, but "tile confirmations" are like 1/15 Hz.

So instead of "wait 6 confirmations for a tx to be secure" it's like "wait D tile confirmations to make sure no double-spend attempt can trick you", where D is the distance between the tiles that you're interacting between.

this is another asymmetry flip like in DoS and DAGs section.